cloud_services/requirements.txt
Flask==2.3.3
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0
pandas==2.0.3
gunicorn==21.2.0